Move `FileHandle` extension implementation to its own file
